+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Some fixes and extra debugging for Update to DDB functionality.
 * Some updates did not work with certain midi-effects.
 * If creating Midi-QOL effects on monsters, will now generate Legendary Resistance effect.
+* ECharacter Effect: Indomitable
 
 # 2.9.53